ADM-348 extract VerifyButton, ResetButton and LoadingDrop components in theme.ts file
