+++ frontend/src/theme.ts
@@ -1,6 +1,7 @@
-import { createTheme } from '@mui/material/styles'
+import { createTheme, styled } from '@mui/material/styles'
 import { indigo } from '@mui/material/colors'
 import { FIVE_HUNDRED } from '@src/constants'
+import { Backdrop, Button } from '@mui/material'
 
 declare module '@mui/material/styles' {
   interface Theme {
@@ -39,4 +40,23 @@ export const theme = createTheme({
       textTransform: 'none',
     },
   },
+})
+
+export const LoadingDrop = styled(Backdrop)({
+  position: 'absolute',
+  zIndex: '999',
+  backgroundColor: 'rgba(199,199,199,0.43)',
+  color: theme.main.backgroundColor,
+})
+
+export const VerifyButton = styled(Button)({
+  width: '3rem',
+  fontSize: '0.8rem',
+  fontWeight: '550',
+})
+export const ResetButton = styled(Button)({
+  width: '3rem',
+  fontSize: '0.8rem',
+  fontWeight: '550',
+  color: '#f44336',
 })
